requirements.txt
jupyter
ipykernel
PyYAML==6.0.1
rasterio==1.3.8
google-cloud-storage==2.16.0
geopandas==1.0.1
earthengine-api==1.1.5
osmnx==1.6.0
networkx==3.3
pandas==2.0.3
numpy==1.25.2
contextily ==1.5.0
shapely==2.0.4
matplotlib==3.7.2
google-api-core==2.19.0
requests==2.31.0
boto3==1.28.37
# pyquadkey2==0.2.2 
scipy==1.11.2
xarray==2023.8.0
fiona==1.10.1
geopy==2.4.1
google-cloud-firestore==2.20.0
google-cloud-run==0.10.14
seaborn==0.12.2
plotly==5.17.0
adjustText==1.3.0
beautifulsoup4==4.12.2
lxml==5.3.0
squarify==0.4.3
tqdm==4.66.1